requirements.txt
requests
requests-html
python-dotenv